+++ packages/contracts/src/http/route.ts
@@ -1,5 +1,9 @@
 'use strict'
 
+import { HttpContext } from './context'
+import { HttpMethods } from './methods'
+import { RouteHandler } from './router'
+
 export interface HttpRoute {
   /**
    * Assign a route prefix.
@@ -10,4 +14,24 @@ export interface HttpRoute {
    * Assign a middleware to the route.
    */
   middleware(middleware: string | string[]): this
+
+  /**
+   * Returns the route path.
+   */
+  path (): string
+
+  /**
+   * Returns the route’s HTTP methods.
+   */
+  methods (): HttpMethods[]
+
+  /**
+   * Returns the route handler.
+   */
+  handler (): RouteHandler
+
+  /**
+   * Run the route handler.
+   */
+  run (ctx: HttpContext): Promise<any>
 }